ENH: reduce speed options when enable auto slow down
Because of extrusion issue under different speed, the printer
is requested to not change speed too frequently.

So we reduce speed options when enable auto slow down

+++ src/libslic3r/PerimeterGenerator.hpp
@@ -29,7 +29,7 @@ public:
     ExtrusionEntityCollection   *gap_fill;
     SurfaceCollection           *fill_surfaces;
 
-    std::map<float, Polygons> m_lower_polygons_series;
+    std::map<int, Polygons> m_lower_polygons_series;
     ExPolygons fill_no_overlap;
     
     PerimeterGenerator(
